feat: 콘서트 단건 조회 api 작성 (#3)
/api/v1/concerts/{concertId} api 작성
단순히 콘서트를 단건 조회하는 로직

+++ src/main/java/com/example/ticketing/domain/concert/dto/response/ConcertRankResponse.java
@@ -5,13 +5,13 @@ import lombok.Getter;
 
 @Getter
 public class ConcertRankResponse {
-    private final Long rank;
+    private final Integer rank;
     private final Long viewCount;
     private final Long concertId;
     private final String concertName;
 
     @Builder
-    public ConcertRankResponse(Long rank, Long viewCount, Long concertId, String concertName){
+    public ConcertRankResponse(Integer rank, Long viewCount, Long concertId, String concertName){
         this.rank = rank;
         this.viewCount = viewCount;
         this.concertId = concertId;
